+++ spinifex/testutil/vbscan/vbscan.go
@@ -12,7 +12,7 @@
 // reimplemented locally rather than imported: spinifex cannot depend on
 // mulga, the module that contains it, without a repo-level cycle. It is
 // deliberately its own leaf package, not part of the shared spinifex/testutil
-// package that ~80 test files import — see testutil/predastore's package
+// package that ~80 test files import — see tests/fixtures/predastore's package
 // comment for why that matters (goleak poisoning via a transitively-linked
 // init-time goroutine). vbscan does not import predastore/quic/quicclient,
 // directly or transitively, and must stay that way.
